+++ packages/apply-release-plan/src/test-utils/get-changelog-entry-with-git-hash.ts
@@ -7,7 +7,7 @@ import { capitalize } from "../utils.ts";
 async function getReleaseLine(changeset: NewChangeset, cwd: string) {
   const [firstLine, ...futureLines] = changeset.summary
     .split("\n")
-    .map((l) => l.trimRight());
+    .map((l) => l.trimEnd());
 
   const [commitThatAddsFile] = await getCommitsThatAddFiles(
     [`.changeset/${changeset.id}.md`],