+++ packages/cli/src/index.ts
@@ -16,6 +16,7 @@ const parsed = mri(args, {
     "otp",
     "since",
     "ignore",
+    "message",
     "tag",
     "snapshot",
     "snapshotPrereleaseTemplate",
@@ -24,6 +25,7 @@ const parsed = mri(args, {
     // Short flags
     v: "verbose",
     o: "output",
+    m: "message",
     // Support kebab-case flags
     "git-tag": "gitTag",
     "snapshot-prerelease-template": "snapshotPrereleaseTemplate",
@@ -50,7 +52,7 @@ if (parsed.help && args.length === 1) {
     $ changeset [command]
   Commands
     init
-    add [--empty] [--open]
+    add [--empty] [--open] [--since <branch>] [--message <text>]
     version [--ignore] [--snapshot <?name>] [--snapshot-prerelease-template <template>]
     publish [--tag <name>] [--otp <code>] [--no-git-tag]
     status [--since <branch>] [--verbose] [--output JSON_FILE.json]