+++ packages/write/src/index.ts
@@ -20,10 +20,10 @@ function getPrettierInstance(cwd: string): typeof prettier {
 
 async function writeChangeset(
   { summary, releases }: Changeset,
-  cwd: string,
+  rootDir: string,
   options?: { prettier?: boolean }
 ): Promise<string> {
-  const changesetBase = path.resolve(cwd, ".changeset");
+  const changesetBase = path.resolve(rootDir, ".changeset");
 
   // Worth understanding that the ID merely needs to be a unique hash to avoid git conflicts
   // experimenting with human readable ids to make finding changesets easier
@@ -33,7 +33,7 @@ async function writeChangeset(
   });
 
   const prettierInstance =
-    options?.prettier !== false ? getPrettierInstance(cwd) : undefined;
+    options?.prettier !== false ? getPrettierInstance(rootDir) : undefined;
   const newChangesetPath = path.resolve(changesetBase, `${changesetID}.md`);
 
   // NOTE: The quotation marks in here are really important even though they are